Split out morph and morph animation, storing data in EGMData
