+++ io_scene_nif/modules/geometry/morph/morph_export.py
@@ -36,21 +36,36 @@
 # POSSIBILITY OF SUCH DAMAGE.
 #
 # ***** END LICENSE BLOCK *****
-import os
 
-import bpy
 from pyffi.formats.egm import EgmFormat
-from pyffi.formats.nif import NifFormat
 
-from io_scene_nif.modules.obj.block_registry import block_store
-from io_scene_nif.utility.util_global import NifOp
+from io_scene_nif.modules.animation.morph_export import MorphAnimation
+from io_scene_nif.utility.util_global import EGMData
 from io_scene_nif.utility.util_logging import NifLog
 
 
-class MorphHelper:
+class Morph:
 
     def __init__(self):
-        self.egm_data = None
+        EGMData.data = None
+        self.morph_animation = MorphAnimation()
+
+    def export_egm(self, key_blocks):
+        EGMData.data = EgmFormat.Data(num_vertices=len(key_blocks[0].data))
+        for key_block in key_blocks:
+            if key_block.name.startswith("EGM SYM"):
+                morph = EGMData.data.add_sym_morph()
+            elif key_block.name.startswith("EGM ASYM"):
+                morph = EGMData.data.add_asym_morph()
+            else:
+                continue
+            NifLog.info("Exporting morph {0} to egm".format(key_block.name))
+            relative_vertices = []
+
+            # note: key_blocks[0] is base key
+            for vert, key_vert in zip(key_blocks[0].data, key_block.data):
+                relative_vertices.append(key_vert - vert)
+            morph.set_relative_vertices(relative_vertices)
 
     def export_morph(self, b_mesh, b_obj, tridata, trishape, vertlist, vertmap):
         # shape key morphing
@@ -64,134 +79,4 @@ class MorphHelper:
                     self.export_egm(key.key_blocks)
                 # TODO [animation] pull out to
                 elif key.ipo:
-                    # regular morph_data export
-                    # (there must be a shape ipo)
-                    keyipo = key.ipo
-                    # check that they are relative shape keys
-                    if not key.relative:
-                        # XXX if we do "key.relative = True"
-                        # XXX would this automatically fix the keys?
-                        raise ValueError("Can only export relative shape keys.")
-
-                    # create geometry morph controller
-                    morph_ctrl = block_store.create_block("NiGeomMorpherController", keyipo)
-                    morph_ctrl.target = trishape
-                    morph_ctrl.frequency = 1.0
-                    morph_ctrl.phase = 0.0
-                    trishape.add_controller(morph_ctrl)
-                    ctrl_start = 1000000.0
-                    ctrl_stop = -1000000.0
-                    ctrl_flags = 0x000c
-
-                    # create geometry morph data
-                    morph_data = block_store.create_block("NiMorphData", keyipo)
-                    morph_ctrl.data = morph_data
-                    morph_data.num_morphs = len(key.key_blocks)
-                    morph_data.num_vertices = len(vertlist)
-                    morph_data.morphs.update_size()
-
-                    # create interpolators (for newer nif versions)
-                    morph_ctrl.num_interpolators = len(key.key_blocks)
-                    morph_ctrl.interpolators.update_size()
-
-                    # interpolator weights (for Fallout 3)
-                    morph_ctrl.interpolator_weights.update_size()
-
-                    # TODO [morph] some unknowns, bethesda only
-                    # TODO [morph] just guessing here, data seems to be zero always
-                    morph_ctrl.num_unknown_ints = len(key.key_blocks)
-                    morph_ctrl.unknown_ints.update_size()
-
-                    for keyblocknum, keyblock in enumerate(key.key_blocks):
-                        # export morphed vertices
-                        morph = morph_data.morphs[keyblocknum]
-                        morph.frame_name = keyblock.name
-                        NifLog.info("Exporting morph {0}: vertices".format(keyblock.name))
-                        morph.arg = morph_data.num_vertices
-                        morph.vectors.update_size()
-                        for b_v_index, (vert_indices, vert) in enumerate(list(zip(vertmap, keyblock.data))):
-                            # vertmap check
-                            if not vert_indices:
-                                continue
-                            # copy vertex and assign morph vertex
-                            mv = vert.copy()
-                            if keyblocknum > 0:
-                                mv.x -= b_mesh.vertices[b_v_index].co.x
-                                mv.y -= b_mesh.vertices[b_v_index].co.y
-                                mv.z -= b_mesh.vertices[b_v_index].co.z
-                            for vert_index in vert_indices:
-                                morph.vectors[vert_index].x = mv.x
-                                morph.vectors[vert_index].y = mv.y
-                                morph.vectors[vert_index].z = mv.z
-
-                        # export ipo shape key curve
-                        curve = keyipo[keyblock.name]
-
-                        # create interpolator for shape key (needs to be there even if there is no curve)
-                        interpol = block_store.create_block("NiFloatInterpolator")
-                        interpol.value = 0
-                        morph_ctrl.interpolators[keyblocknum] = interpol
-                        # fallout 3 stores interpolators inside the interpolator_weights block
-                        morph_ctrl.interpolator_weights[keyblocknum].interpolator = interpol
-
-                        # geometry only export has no float data also skip keys that have no curve (such as base key)
-                        if NifOp.props.animation == 'GEOM_NIF' or not curve:
-                            continue
-
-                        # note: we set data on morph for older nifs and on floatdata for newer nifs
-                        # of course only one of these will be actually written to the file
-                        NifLog.info("Exporting morph {0}: curve".format(keyblock.name))
-                        interpol.data = block_store.create_block("NiFloatData", curve)
-                        floatdata = interpol.data.data
-                        if curve.getExtrapolation() == "Constant":
-                            ctrl_flags = 0x000c
-                        elif curve.getExtrapolation() == "Cyclic":
-                            ctrl_flags = 0x0008
-
-                        morph.interpolation = NifFormat.KeyType.LINEAR_KEY
-                        morph.num_keys = len(curve.getPoints())
-                        morph.keys.update_size()
-
-                        floatdata.interpolation = NifFormat.KeyType.LINEAR_KEY
-                        floatdata.num_keys = len(curve.getPoints())
-                        floatdata.keys.update_size()
-
-                        start = bpy.context.scene.frame_start
-                        fps = NifOp.context.scene.render.fps
-                        for i, btriple in enumerate(curve.getPoints()):
-                            knot = btriple.getPoints()
-                            morph.keys[i].arg = morph.interpolation
-                            morph.keys[i].time = (knot[0] - start) * fps
-                            morph.keys[i].value = curve.evaluate(knot[0])
-                            # morph.keys[i].forwardTangent = 0.0 # ?
-                            # morph.keys[i].backwardTangent = 0.0 # ?
-                            floatdata.keys[i].arg = floatdata.interpolation
-                            floatdata.keys[i].time = (knot[0] - start) * fps
-                            floatdata.keys[i].value = curve.evaluate(knot[0])
-                            # floatdata.keys[i].forwardTangent = 0.0 # ?
-                            # floatdata.keys[i].backwardTangent = 0.0 # ?
-                            ctrl_start = min(ctrl_start, morph.keys[i].time)
-                            ctrl_stop = max(ctrl_stop, morph.keys[i].time)
-                    morph_ctrl.flags = ctrl_flags
-                    morph_ctrl.start_time = ctrl_start
-                    morph_ctrl.stop_time = ctrl_stop
-
-                    # fix data consistency type
-                    tridata.consistency_flags = b_obj.niftools.consistency_flags
-
-    def export_egm(self, key_blocks):
-        self.egm_data = EgmFormat.Data(num_vertices=len(key_blocks[0].data))
-        for key_block in key_blocks:
-            if key_block.name.startswith("EGM SYM"):
-                morph = self.egm_data.add_sym_morph()
-            elif key_block.name.startswith("EGM ASYM"):
-                morph = self.egm_data.add_asym_morph()
-            else:
-                continue
-            NifLog.info("Exporting morph %s to egm" % key_block.name)
-            relative_vertices = []
-
-            # note: key_blocks[0] is base key
-            for vert, key_vert in zip(key_blocks[0].data, key_block.data):
-                relative_vertices.append(key_vert - vert)
-            morph.set_relative_vertices(relative_vertices)
+                    self.morph_animation.export_morph_animation(b_mesh, b_obj, key, tridata, trishape, vertlist, vertmap)
