+++ io_scene_nif/utility/util_global.py
@@ -75,3 +75,11 @@ class KFData:
 
     def __init__(self):
         pass
+
+
+class EGMData:
+
+    data = None
+
+    def __init__(self):
+        pass
